requirements.txt
appdirs==1.4.4
asgiref==3.3.4
Babel==2.9.0
backcall==0.2.0
black==21.4b0
cfgv==3.2.0
click==7.1.2
dataclasses==0.6
decorator==5.0.7
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.14
django-crispy-forms==1.11.2
django-extra-views==0.13.0
django-haystack==3.0
django-heroku==0.3.1
django-import-export==2.5.0
django-oscar==2.1
django-oscar-accounts==2.1
django-phonenumber-field==3.0.1
django-tables2==2.2.1
django-treebeard==4.5.1
django-widget-tweaks==1.4.8
et-xmlfile==1.1.0
factory-boy==2.12.0
Faker==8.1.1
filelock==3.0.12
flake8==3.9.1
gunicorn==20.1.0
identify==2.2.4
importlib-metadata==4.0.1
importlib-resources==5.1.2
ipdb==0.13.7
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
MarkupPy==1.14
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
odfpy==1.4.1
openpyxl==3.0.7
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
phonenumbers==8.12.21
pickleshare==0.7.5
Pillow==8.2.0
pre-commit==2.12.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
purl==1.5
pycodestyle==2.7.0
pycountry==20.7.3
pyflakes==2.3.1
Pygments==2.8.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
tablib==3.0.0
text-unidecode==1.3
toml==0.10.2
traitlets==4.3.3
typed-ast==1.4.3
typing-extensions==3.7.4.3
virtualenv==20.4.4
wcwidth==0.2.5
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
